+++ frontend/src/app/modules/page/admin/core/admin.model.ts
@@ -20,15 +20,19 @@
  ********************************************************************************/
 
 
-import {FormArray, FormControl, FormGroup} from '@angular/forms';
-import {CalendarDateModel} from '@core/model/calendar-date.model';
-import {Pagination, PaginationResponse} from '@core/model/pagination.model';
+import { FormArray, FormControl, FormGroup } from '@angular/forms';
+import { CalendarDateModel } from '@core/model/calendar-date.model';
+import { Pagination, PaginationResponse } from '@core/model/pagination.model';
 
 export enum KnownAdminRoutes {
   BPN = 'configure-bpn',
   IMPORT = 'configure-import',
   CONTRACT = 'contracts',
-  CONTRACT_DETAIL_VIEW = 'contracts/:contractId'
+  CONTRACT_DETAIL_VIEW = 'contracts/:contractId',
+  POLICY_MANAGEMENT = 'policies',
+  POLICY_MANAGEMENT_EDIT = 'policies/edit/:policyId',
+  POLICY_MANAGEMENT_CREATE = 'policies/create',
+  POLICY_MANAGEMENT_DETAIL_VIEW = 'policies/:policyId',
 }